+++ R/contrastiveCov.R
@@ -11,15 +11,26 @@
 #'  data sets should be centered to mean zero.
 #' @param scale A \code{logical} indicating whether the target and background
 #'  data sets should be scaled to unit variance.
+#' @param scaled_matrix A \code{logical} indicating whether to output a
+#'  \code{\link[ScaledMatrix]{ScaledMatrix}} object. The centering and scaling
+#'  procedure is delayed until later, permitting more efficient matrix
+#'  multiplication and row or column sums downstream. However, this comes at the
+#'  at the cost of numerical precision. Defaults to \code{FALSE}.
 #'
 #' @return A list of contrastive covariance matrices. Each element has an
 #'   associated contrastive parameter in the \code{contrasts} vector.
 #'
 #' @keywords internal
-contrastiveCov <- function(target, background, contrasts, center, scale) {
+contrastiveCov <- function(
+  target, background, contrasts, center, scale, scaled_matrix = FALSE
+) {
   # get the covariance matrices of the target and background
-  c_target <- covMat(target, center = center, scale = scale)
-  c_background <- covMat(background, center = center, scale = scale)
+  c_target <- covMat(
+    target, center = center, scale = scale, scaled_matrix = scaled_matrix
+  )
+  c_background <- covMat(
+    background, center = center, scale = scale, scaled_matrix = scaled_matrix
+  )
 
   # get the list of contrastive covariance matrices
   c_contrasts <- lapply(contrasts, function(x) {
@@ -46,6 +57,11 @@ contrastiveCov <- function(target, background, contrasts, center, scale) {
 #'  data sets should be centered to mean zero.
 #' @param scale A \code{logical} indicating whether the target and background
 #'  data sets should be scaled to unit variance.
+#' @param scaled_matrix A \code{logical} indicating whether to output a
+#'  \code{\link[ScaledMatrix]{ScaledMatrix}} object. The centering and scaling
+#'  procedure is delayed until later, permitting more efficient matrix
+#'  multiplication and row or column sums downstream. However, this comes at the
+#'  at the cost of numerical precision. Defaults to \code{FALSE}.
 #'
 #' @return A list of contrastive covariance matrices. Each element has an
 #'   associated contrastive parameter in the \code{contrasts} vector.
@@ -53,10 +69,17 @@ contrastiveCov <- function(target, background, contrasts, center, scale) {
 #' @importFrom BiocParallel bplapply
 #'
 #' @keywords internal
-bpContrastiveCov <- function(target, background, contrasts, center, scale) {
+bpContrastiveCov <- function(
+  target, background, contrasts, center, scale, scaled_matrix = FALSE
+) {
+
   # get the covariance matrices of the target and background
-  c_target <- covMat(target, center = center, scale = scale)
-  c_background <- covMat(background, center = center, scale = scale)
+  c_target <- covMat(
+    target, center = center, scale = scale, scaled_matrix = scaled_matrix
+  )
+  c_background <- covMat(
+    background, center = center, scale = scale, scaled_matrix = scaled_matrix
+  )
 
   # get the list of contrastive covariance matrices
   c_contrasts <- BiocParallel::bplapply(contrasts, function(x) {